Move SuscriptionMailer into Suscribir module.
